src/app/Service/User Location/weather.service.ts
import { Injectable } from '@angular/core';
import { HttpClient } from '@angular/common/http';
import { Observable } from 'rxjs';


@Injectable({
  providedIn: 'root'
})
export class WeatherService {

  private apiKey = '7cfc7aa1e9a7992ae745363e9f342848';
  private apiUrl = 'https://api.openweathermap.org/data/2.5/weather';

  constructor(private http: HttpClient) {}

  getWeatherByLocation(lat: number, lon: number) {
    const params = {
      lat: lat.toString(),
      lon: lon.toString(),
      appid: this.apiKey,
      units: 'metric' 
    };

    return this.http.get(`${this.apiUrl}`, { params });
}

}